+++ mock/mock_client.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -source=./cloud/scope/client.go -destination ./cloud/scope/mock/client.go -package mock
+//	mockgen -source=./cloud/scope/client.go -destination ./mock/mock_client.go -package mockobjects
 //
 
-// Package mock is a generated GoMock package.
-package mock
+// Package mockobjects is a generated GoMock package.
+package mockobjects
 
 import (
 	context "context"
